+++ spring-cloud-tencent-polaris-context/src/main/java/com/tencent/cloud/polaris/context/config/PolarisContextAutoConfiguration.java
@@ -16,19 +16,17 @@
  *
  */
 
-package com.tencent.cloud.polaris.context;
+package com.tencent.cloud.polaris.context.config;
 
 import java.util.List;
 
-import com.tencent.cloud.common.metadata.StaticMetadataManager;
-import com.tencent.cloud.common.metadata.config.MetadataAutoConfiguration;
+import com.tencent.cloud.polaris.context.ConditionalOnPolarisEnabled;
+import com.tencent.cloud.polaris.context.ModifyAddress;
+import com.tencent.cloud.polaris.context.PolarisConfigModifier;
+import com.tencent.cloud.polaris.context.ServiceRuleManager;
 import com.tencent.polaris.api.exception.PolarisException;
-import com.tencent.polaris.api.plugin.common.ValueContext;
-import com.tencent.polaris.api.plugin.route.LocationLevel;
 import com.tencent.polaris.client.api.SDKContext;
-import org.apache.commons.lang.StringUtils;
 
-import org.springframework.boot.autoconfigure.ImportAutoConfiguration;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.context.annotation.Bean;
@@ -41,33 +39,14 @@ import org.springframework.core.env.Environment;
  */
 @ConditionalOnPolarisEnabled
 @EnableConfigurationProperties({PolarisContextProperties.class})
-@ImportAutoConfiguration(MetadataAutoConfiguration.class)
 public class PolarisContextAutoConfiguration {
 
 	@Bean(name = "polarisContext", initMethod = "init", destroyMethod = "destroy")
 	@ConditionalOnMissingBean
 	public SDKContext polarisContext(PolarisContextProperties properties, Environment environment,
-			List<PolarisConfigModifier> modifierList, StaticMetadataManager staticMetadataManager)
+			List<PolarisConfigModifier> modifierList)
 			throws PolarisException {
-		SDKContext sdkContext = SDKContext.initContextByConfig(properties.configuration(environment, modifierList));
-
-		// init current instance location info from environment
-		ValueContext valueContext = sdkContext.getValueContext();
-		String region = staticMetadataManager.getRegion();
-		String zone = staticMetadataManager.getZone();
-		String campus = staticMetadataManager.getCampus();
-
-		if (StringUtils.isNotBlank(region)) {
-			valueContext.setValue(LocationLevel.region.name(), region);
-		}
-		if (StringUtils.isNotBlank(zone)) {
-			valueContext.setValue(LocationLevel.zone.name(), zone);
-		}
-		if (StringUtils.isNotBlank(campus)) {
-			valueContext.setValue(LocationLevel.campus.name(), campus);
-		}
-
-		return sdkContext;
+		return SDKContext.initContextByConfig(properties.configuration(environment, modifierList));
 	}
 
 	@Bean